requirements.txt
discord.py
python-dotenv
aiosqlite
together
asyncio